+++ packages/reactivity-core/struct/struct.ts
@@ -85,7 +85,9 @@ type AnyFunc = (...args: any[]) => any;
  * @param T The type of the struct.
  * @param V The value of the property.
  */
-type GetMemberSchemaForProp<T, V> = V extends AnyFunc
+type GetMemberSchemaForProp<T, V> = [V] extends [AnyFunc]
+    // V is wrapped in an array to prevent distributive conditional types
+    // see https://www.typescriptlang.org/docs/handbook/2/conditional-types.html#distributive-conditional-types
     ?
           | PropertyMemberType
           | MethodMemberType<T, Parameters<V>, ReturnType<V>>